refactor: extract GlobalChangeOperation and PartitionGroupOperation into separate files
Move all operation records out of ClusterConfigurationChangeOperation.java
into two new top-level sealed sub-interface files that mirror the proto
topology.proto split:

- GlobalChangeOperation.java: broker lifecycle ops (MemberJoin/Leave/Remove,
  PreScaling, PostScaling, UpdatePartitionDistributorConfig)
- PartitionGroupOperation.java: partition-group ops (DeleteHistory,
  UpdateRoutingState, UpdateIncarnationNumber, ModeChange, AwaitModeChange,
  ScaleUpOperation, PartitionChangeOperation)

ClusterConfigurationChangeOperation.java is now a thin sealed interface
that only declares memberId() and permits the two sub-interfaces.

Narrow the phase list types in PhasedChangePlan: GlobalPhase now carries
List<GlobalChangeOperation> and PartitionGroupParallelPhase carries
Map<String, List<PartitionGroupOperation>>, replacing the previous
List<ClusterConfigurationChangeOperation>, so callers cannot accidentally
mix operation categories at the type level.

All callers updated to use the new import paths.
